+++ CHANGELOG.md
@@ -235,6 +235,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
   - From\<T: PrimInt\> for Der (`PrimInt` is `i8` or `u8` or ... `u128` or `isize` or `usize`.)
   - From\<&str\> for Der
   - From\<&[u8]\> for Der
+  - From\<&ContentsRef\> for Contents
   - From\<&[u8; N]\> for Contents
   - From\<&[u8; N]\> for &ContentsRef
   - From\<&mut [u8; N]\> for &mut ContentsRef
@@ -268,6 +269,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
   - Der::octet_string()
   - Der::utf8_string()
   - Contents::from\_bool()
+  - Contents::from\_bytes()
   - ContentsRef::from\_bool()
   - ContentsRef::from\_bytes()
   - ContentsRef::from\_mut\_bytes()